Renamed a method; Transferred some methods to Util.java; Some other minor changes
